requirements.txt
google-api-python-client==2.80.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
mysql-connector-python==8.0.33
gitpython==3.1.31
